Extract file works for directories
Extract file requests will now check the filesystem to see if that path
exists in the package.  If not, and the package can be extracted, it will
fallback to the previous behaviour of extracting a file with the given path
and returning that.

+++ storage_service/locations/api/resources.py
@@ -362,48 +362,59 @@ class PackageResource(ModelResource):
             mimetype='application/json')
 
     def extract_file_request(self, request, **kwargs):
+        """
+        Returns a single file from the Package, extracting if necessary.
+        """
         relative_path_to_file = request.GET.get('relative_path_to_file')
+        temp_dir = ''
 
         # Tastypie checks
         self.method_check(request, allowed=['get'])
         self.is_authenticated(request)
         self.throttle_check(request)
 
-        # Get AIP details
+        # Get Package details
         package = Package.objects.get(uuid=kwargs['uuid'])
-        if package.package_type not in Package.PACKAGE_TYPE_EXTRACTABLE:
-            # Can only extract files from AIPs
-            return http.HttpMethodNotAllowed()
+        full_path = package.full_path()
 
-        # create temp dir to extract to
-        temp_dir = tempfile.mkdtemp()
+        local_path = os.path.join(full_path, relative_path_to_file)
+        if os.path.exists(local_path):
+            # Local file exists - return that
+            extracted_file_path = local_path
+        elif package.package_type in Package.PACKAGE_TYPE_EXTRACTABLE:
+            # If file doesn't exist, try to extract it
 
-        filename, file_extension = os.path.splitext(package.full_path())
+            # Create temp dir to extract to
+            # TODO store this somewhere well known, so we don't clutter up
+            # filesystem with duplicates of files?
+            temp_dir = tempfile.mkdtemp()
 
-        # extract file from AIP
-        if file_extension == '.bz2':
-            command_data = [
-                'tar',
-                'xvjf',
-                package.full_path(),
-                '-C' + temp_dir,
-                relative_path_to_file
-            ]
-        else:
-            command_data = [
-                '7za',
-                'e',
-                '-o' + temp_dir,
-                package.full_path(),
-                relative_path_to_file
-            ]
+            filename, file_extension = os.path.splitext(full_path)
 
-        subprocess.call(command_data)
-        # send extracted file
-        if file_extension == '.bz2':
-            extracted_file_path = os.path.join(temp_dir, relative_path_to_file)
-        else:
-            extracted_file_path = os.path.join(temp_dir, os.path.basename(relative_path_to_file))
+            # extract file from Package
+            if file_extension == '.bz2':
+                command_data = [
+                    'tar',
+                    'xvjf',
+                    full_path,
+                    '-C' + temp_dir,
+                    relative_path_to_file
+                ]
+            else:
+                command_data = [
+                    '7za',
+                    'e',
+                    '-o' + temp_dir,
+                    full_path,
+                    relative_path_to_file
+                ]
+
+            subprocess.call(command_data)
+            # send extracted file
+            if file_extension == '.bz2':
+                extracted_file_path = os.path.join(temp_dir, relative_path_to_file)
+            else:
+                extracted_file_path = os.path.join(temp_dir, os.path.basename(relative_path_to_file))
 
         # If not found, return 404
         if not os.path.exists(extracted_file_path):
@@ -427,12 +438,17 @@ class PackageResource(ModelResource):
 
         response['Content-Length'] = os.path.getsize(extracted_file_path)
 
-        self.log_throttled_access(request)
+        # Delete temp dir if created
+        if os.path.exists(temp_dir):
+            shutil.rmtree(temp_dir)
 
-        shutil.rmtree(temp_dir)
+        self.log_throttled_access(request)
         return response
 
     def download_request(self, request, **kwargs):
+        """
+        Returns the entire Package to be downloaded.
+        """
         # Tastypie checks
         self.method_check(request, allowed=['get'])
         self.is_authenticated(request)
@@ -441,7 +457,8 @@ class PackageResource(ModelResource):
         # Get AIP details
         package = Package.objects.get(uuid=kwargs['uuid'])
         if package.package_type not in Package.PACKAGE_TYPE_EXTRACTABLE:
-            # Can only extract files from AIPs
+            # Can only return packages that are a single file
+            # TODO Update to zip up a transfer before returning it?
             return http.HttpMethodNotAllowed()
 
         filename = os.path.basename(package.full_path())
